New structure for the `verifier` pkg
which now looks like:
```
./pkg/kudoctl/packages/verifier/
├── template
│   ├── parser.go
│   ├── parser_test.go
│   ├── verify_parameters.go
│   ├── verify_parameters_test.go.go
│   ├── verify_references.go
│   └── verify_references_test.go
└── types.go
```

- moved `templates` pkg into `verify/templates` - this is the subpkg that handles verifying templates, leaving it open to verify other things (like operator.yaml etc.)
- split existing verifiers into separate files
- renamed `template_parsing` to `parser` since _template_ is already in the package name
